Make FileInStream an abstract class.
