+++ job/server/src/test/java/alluxio/job/plan/migrate/MigrateDefinitionRunTaskTest.java
@@ -91,7 +91,7 @@ public final class MigrateDefinitionRunTaskTest {
         .thenReturn(ClientContext.create(conf));
     when(mMockFileSystemContext.getClusterConf()).thenReturn(conf);
     when(mMockFileSystemContext.getPathConf(any(AlluxioURI.class))).thenReturn(conf);
-    mMockInStream = new MockFileInStream(mMockFileSystemContext, TEST_SOURCE_CONTENTS, conf);
+    mMockInStream = new MockFileInStream(TEST_SOURCE_CONTENTS);
     when(mMockFileSystem.openFile(eq(new AlluxioURI(TEST_SOURCE)),
         any(OpenFilePOptions.class))).thenReturn(mMockInStream);
     mMockOutStream = new MockFileOutStream(mMockFileSystemContext);
